Modified partner email compromise to generate compromsied partner senders
